* -skip replaces -tempskip; new version also skips SSA updates.  Seems to test well for PST experiments.
* I've done some testing with norm choice for viscosity convergence
of SSA outer iteration.  Using PST experiments.  Conclude that L^1 is
best because it gives same result for verification tests (sans margins)
but does better for PST (with margins) because it appropriately ignors
details of changes in viscosity at those few margin locations.

* Revised documentation to reflect above testing and changes.

* Fixed examples/eisgreen/eis_core.py to use Python more effectively
and to write an UNLIMITED time dimension.

* Added Antarctica related references to doc/ice_bib.bib.