Make Ross work again (bug #13072)
Put IceType::hardnessParameter back, this is not entirely desirable, but
we'll live with it for now.

Use pressure-adjusted temperature when inverting Sigma.

Respect the enhancement factor when inverting Sigma.  What is the right
thing to do here, enhancement factor is not used in SSA.  We should
probably get rid of it, but that's for another time.

Note that the reason output is not identical to the "normal output"
attachment for bug #13072 is that we're using pressure-adjusted
temperature now.