Fix of bug #13354 in stable0.2.  Sign fix, as suggested by Ziemann; sorry ...
Note  oceanPCC->writeCouplingFieldstoFile()  was not called at
all and now it is.  But PISMConstOceanCoupler defaults to no write to maintain consistent 
stable0.2 behavior; PISMOceanCoupler and derived classes will, presumably, write nontrivial
fields for shelf bases.

Clarification in comments on why sign is the way it is: agrees with grounded basal melt
rate.