Replacement of old couplers (PDDCoupler, MonthlyTempsCoupler, and the PDD in EISMINT-Greenland by the new system (SnowModelCoupler + LocalMassBalance and MonthlyDataMaps classes) largely complete.  User's Manual not updated yet.
IceGRNModel uses the new system.  It is a model citizen.  (It would be
reasonable to put the Ellesmere Island delete stuff into a python script,
by the way.)

The basic issue is that there are two surface temperatures, namely the one
used as the upper boundary condition for the conservation of energy in the 
ice (ice surface temp = artm) and the one used by the mass balance scheme to convert
snow fall rate (snowprecip) to a signed mass flux into the ice.  The latter
temperature is the snow-surface temperature 'snowtemp' in the output file.
Fausto et al, and the EISMINT-Greenland spec, give parameterizations for 
the latter.

In particular, at initialization it is not clear if when we see 'artm' which
it means, and if we have a parameterization, whether to use it for the ice
surface temp.