1. Refactor of main derived class of PISMAtmosphereCoupler: - namely, PISMSnowModelAtmosCoupler is renamed PISMGreenlandAtmosCoupler - it is put into separate file src/coupler/pGreenlandAtmosCoupler.[hh|cc] - it has the SeaRISE-chosen Fausto et al (2009) parameterization for temperature,   as before - EISGREENAtmosCoupler is still a derived class of it - monthly temperature inputs presumably broken (but that mechanism   should change immediately to be SeaRISE future climate forcing case anyway)
2. Change to temperature parameterization: Formerly the temperature used
as the upper boundary condition of conservation of energy was different
from the mean annual snow temperature used in the snow processes (= mass
balance) model.  Now it is the same.  More sophistication is needed, because
what we actually have as input is the 2m air temperature and we need to have
a temperature which is used as the mass balance process temperature (i.e. 
the PDD is computed from it) and we need to have the temperature of the 
ice when it first becomes ice below (at) the end of firn processes.  To do ...

3. Various comments and "using" lines in coupler sources updated. 

4. Options for pcctest changed: -sma now -greenland, to apply PISMGreenlandAtmosCoupler

5. It still remains to implement the SeaRISE paleo-precipitation scheme, which
is given on the wiki.