Cosmetic change, but requires thought and then a bit more coding to fix FIXMEs.
I modified the option name "-anomaly_temp_ma foo.nc" to "-anomaly_temp foo.nc".
This is because my reading of PAForcing::temp_time_series(), called in
PSTemperatureIndex::update(), suggests it uses the time-dependent and
space-dependent temperature anomaly in foo.nc in the "right way" already.
That is, if the time-resolution of the temperature anomaly is finer than a
year then the anomaly is added in the correct way to the temperature time series
seen by the PDD model in determining the amount of melt and rain.

Context: I assert that sub-annual temperature anomalies are more important
climatic forcings through their effect on melt than they are through their
effect on ice fluid upper boundary temperature changes.

Note that the method PAForcing::mean_annual_temp() provides the
temperature anomaly to the surface model, and thus to the ice upper surface
below, through method PSTemperatureIndex::ice_surface_temperature().

Wit this (r1280) commit, PAForcing::mean_annual_temp() does NOT
do the right thing with a sub-annual temperature anomaly.  That is, a sub-annual
resolution anomaly will alter the ice surface upper boundary condition by an
amount which is a *sample* of the combined (mean temperature)+anomaly.  We have
no control over when that sample is taken, and there is no reason to believe
that a sample of the anomaly is the correct mean annual anomaly.

To address the last issue, see the FIXMEs in PAForcing::mean_annual_temp().

Finally there is a broad FIXME:  The variable name temp_ma_anomaly in PAForcing, 
and the "_ma" in the naming schemes generally, can go?