Removed pism_config:water_melting_temperature and replaced with pism_config:water_triple_point_temperature.  Value is unchanged at 273.15.
Removed IceFlowLaw::meltingTemp and replaced with IceFlowLaw::triple_point_temp.

These are trivial changes but they are correct.  They help with updating the
EnthalpyConverter, which will have a reference temperature which is not the
triple point.  Confusion between the triple point and the melting
temperature is not helpful to the EnthalpyConverter issues.

Note that some error checking is now turned off in
GPBLDIce::softnessParameter_from_enth() but apparently it is never active
anyway, in that all regression tests pass with it off.  (I have turned it off in
preparation for an EnthalpyConverter update.)  Why were these checks needed?