First pass at python bindings for (a very, very small part of) PISM.
To build the bindings, you'll need to set the Pism_BUILD_PYTHON_BINDINGS cmake flag.
If you don't do this, nothing should be broken.  If you do do this, you take your chances.

See src/pythonbindings/README.txt for more instructions on building and installing the bindings,
as well as some examples to try if (by some miracle) you are able to build and install the bindings.