Re-implemented the -ocean_kill mechanism and cleaned up some code.
The command-line option -ocean_kill turns on the new code. If an argument (file name) is provided, an ice thickness field is read from this file and is used to set calving front locations. Otherwise we use ice thickness at the beginning of the run, no matter how it was set.

This mechanism removes ice entering areas that are ice-free according to the thickness field mentioned above and are below sea level at the beginning of the run.