Modified the code reading (and using) scalar and spatially-variable forcing data to use time bounds (if present).
Scalar forcing code interprets forcing data as piecewise-constant in time if time bounds are present and as piecewise-linear if they are not present.

Speatially-variable forcing is interpreted as piecewise-constant in time in both cases, at least by default. Option -bc_time_average enables linear interpolation in time (i.e. switches to the piecewise-linear-in-time interpretation).