Fixed bug #18149 in stable0.4.
Now *if* the proj.4 library is present the Pism_USE_PROJ4 option is available. If it is set to "on", proj.4 is used to compute cell areas. If it is set to "off", PISM's old behavior is preserved.

(In other words, if proj.4 is present but the user doesn't ask us to use it PISM does what it did before. PISM stable0.4 did not and still does not *require* proj.4 to build.)