Added the -ssa_dirichlet_bc option to pismo.
It turns on the code that sets SSA Dirichlet B.C. in the no-model strip to values read from an input file.
This code looks for variables ubar_ssa_bc and vbar_ssa_bc.

Also: fixed a bug in iMgeometry.cc (the SSA Dirichlet BC case was not handled correctly).