Fixes the bug (I bet there is a better fix, though).
Also: re-organized a bunch of BedDeformLC code, mostly to make sure I know
exactly what it does.

Here's the issue: when initializing from an uplift map, the computed pre-bent
plate position has to be "tweaked" in a way compatible with the tweaking we do
during time-stepping. Please see the block at the end of
BedDeformLC::uplift_init().