Finishes #19 as I see it.
I believe that the maximum of the horizontal velocity over the volume of the
ice provides the desired information. Moreover, CFL conditions for 2D (mass
continuity) and 3D (enthalpy, age) calculations use somewhat incompatible ways
of computing this "maximum velocity". (I'm confused about the reason why.)

In particular, the CFL for 3D computations uses max(u) and max(v) which are
found independently (i.e. can be achieved in different places), while the
adaptive time-stepping code for the mass continuity computation ties u and v
velocity components.

Only the former "maximums" are easily available for reporting, so right now
PISM reports the magnitude of the vector (max(u), max(v)), which, in general,
exceeds the magnitude of the fastest flow.

Please close this task if this is acceptable or add a comment describing
necessary changes otherwise.