One more PAGivenClimate fix.
air_temp should also be averaged in time so that PAGivenClimate reports correct mean annual air temperature, which is used as the top-of-the-ice temperature. This does the "right thing" in the "-surface pdd -pdd_annualize -atmosphere given" case. In many other cases (i.e. when PAGivenClimate takes sub-one-year time steps) mean_annual_temp cannot be computed anyway, so this change does not matter.