Fixes the PDD code. See #133. Needs testing.
The new code keeps track of the snow depth, but re-sets it to zero
once a year (on pdd_balance_year_start_day, October 1 by default).

For restartability the new field (snow_depth) is saved to output
files. It is not required for re-starting, though: if it is not found
in an input file, snow depth is set to zero.

Both -pdd_annualize and "regular" scheme does what I think they should
do. Using the -pdd_annualize option trades sub-annual ice-dynamics
feedback for a faster computation of the mass balance. In the absence
of such feedback the two schemes are equivalent (assuming that the run
started on October 1).

Also: the new code ignores negative precipitation, which we allowed
for historical reasons (as far as I can tell).