Refactoring.  The base class PISMHydrology holds a Wtil variable because all models need somewhere to put water, and all models need to report what the Mohr-Coulomb relation needs.  The new class PISMNullTransportHydrology updates Wtil by a new (more stable, but easy) implicit method.  This new class can't conserve water and it does not attempt to report conservation.  The "till can" aspect is de-emphasized but still present.
The new names for -hydrology choices are null,routing,distributed.  Till is
always present but it can be turned off with -hydrology_tillwat_max 0.0 or
setting that config constant to 0.

Changes to PISMDistributedHydrology are still needed so that it has correct
coupling W <--> Wtil; at this point it does not compile.