[I am forcing this commit message to appear in pism.pushes by a white-space commit.  Sorry that it ain't elegant.]
Merge till-hydro-dynamics.  The dev branch now has the new hydrology options
  -hydrology null   [default]
  -hydrology routing
  -hydrology distributed
The first two of these are documented in section 6.6 of the User's Manual.
The third is unpublished and undocumented.

The major change for 'dev' branch users is that the field 'tillwat' always
appears in output files, while 'bwat' only appears there with routing or
distributed models.

For most purposes in 'dev' and previous versions of PISM, 'tillwat' is the new
name for basal water, replacing the old meaning of 'bwat'.  To use results from
'dev' runs before this commit, rename 'bwat' to 'tillwat' and keep the same
options.

Only for mass-conserving transport hydrologies like routing and distributed is
'bwat' even meaningful (in the new meanings).  In such mass-conserving
hydrologies it refers to the amount of subglacial water moving past the
location, while 'tillwat' refers to the amount bound up in till.