Implement a "constant velocity" stress balance (see #171).
This is a *shallow* stress balance object, i.e. it reads in a 2D
velocity field that can then be used as a "sliding" velocity with the
SIA or with the "trivial" shallow stress balance modifier (this is the
intended use). This way we can set up a run using
constant-in-the-column, time-independent, spatially-variable prescribed
ice velocities.

To enable, use the option -ssb_constant (here "SSB" stands for "shallow
stress balance"). Use -ssb_constant_file to specify a file with
variables "ubar" and "vbar".

By default this will use "ubar" and "vbar" as sliding velocities for the
SIA model. Add "-no_sia" to run with constant-in-the-column velocities.