"Regrid" enthalpy using temp and liqfrac.
This allows re-gridding enthalpy from files that contain temperature and
liquid water fraction, but no enthalpy variable itself.

If both 'temp' and 'liqfrac' are present, use them. If 'liqfrac' is
missing, assume that it is zero and initialize enthalpy as if the ice is
cold. If both 'temp' and 'liqfrac' are missing, stop.