Add -part_grid_reduce_frontal_thickness.
This adds an option that turns on (and off) the code reducing the
threshold ice thickness used by the "part_grid" mechanism.

Here's what this means: When using -part_grid, PISM has to decide if a
"partially-filled" cell is now full. In a flowline setup similar to the
van der Veen analytical profile (in 'Parameterization for subgrid-scale
motion of ice-shelf calving fronts' by Albrecht et al, for example) we
can adjust the threshold thickness using the (spatial) thinning rate
derived using the analytical profile.

While it seems to improve results in this particular setup, it is not
clear if this adjustment should always be used, hence the default is
"off".

See IceModel::get_threshold_thickness() for the implementation.