Address #74 (for now).
This commit documents the fact that sub_shelf_ice_flux may be non-zero
even if iareaf is exactly zero. (Runs with -float_kill are an example.)

This will need a better fix, but documenting this behavior should be
good enough for now.

The script examples/marine/circular/issue-74-test.sh makes it easy to
test if #74 is fixed.