Enthalpy code: try to recover energy conservation at the base.
Ed, Andy, Flo: please take a look at this. There might be a factor of 2
error in there. (At the base the "control volume" corresponding to a
grid point should be half as thick as the regular control volume in the
interior of a column, right? If this is correct, then changing
Enthnew[0] actually changes energy of dx*dy*(1/2*dz_fine) m^3 of ice,
not dx*dy*dz_fine. I'll let you decide.)