More tweaking after reading Holland & Jenkins, 1999.
I realized that both the strong melt case and the freeze-on case *are*
covered by Holland and Jenkins. We don't need *two* models, and we don't
need Hellmer and Olbers at all, because the H+O model is a subset of the
H+J model.

I will add the needed finishing touches tomorrow.