Fix the PDD code in dev. See #254. (Port to 0.6)
The PDD surface mass balance models resets the snow depth once a year,
on the day corresponding to the beginning of the mass balance year. The
code re-setting the snow depth was not designed to work in the case of
non-contiguous updates (i.e. the "cache" modifier). I believe it does
the right thing now.