Create new option and config flag -tauc_add_transportable_water.
This option will work in PISMRoutingHydrology and derived (e.g.
PISMDistributedHydrology).  In PISMNullHydrology it is an error.
It changes the computation of tauc from the water amount.  Instead
of just using tillwat, it uses tillwat+bwat.

Testing of this new feature:  It compiles.  Error messages seem to
occur when appropriate.  In all other ways you are on your own.