Fixing PISM parameter names (Paterson-Budd_*) to remove hyphen, to get CF compliance.
This minor problem arises because pism_config is now included in output files.

The CF convention document excludes hyphens:  "Variable, dimension and attribute
names should begin with a letter and be composed of letters, digits, and
underscores. Note that this is in conformance with the COARDS conventions,
but is more restrictive than the netCDF interface which allows use of the hyphen
character."