Start working on #202.
This commit moves some code from a private repository of mine into PISM.

This code was *not* added to the build system yet.

The ConfigJSON class requires the Jansson
library (https://github.com/akheron/jansson). I will add Jansson to the
PISM source code tree as a Git submodule. (I think we should also fork
it under pism/jansson on GitHub.) Luckily Jansson uses a license that
allows such use and does not conflict with GPL. It also uses CMake as a
build system, so we can integrate it into PISM with very little effort.

The JSON configuration file validator (validate_config.py) requires the
jsonschema Python module (https://pypi.python.org/pypi/jsonschema).

All of this is "alpha-quality" code and change a lot. In particular, it
will use upcoming changes in the PISM error handling code.

The pism_config.json is (approximately) what a PISM's config file will
look like. It is a bit out of date (does not include some parameters
that were added recently) and does not include command-line options
yet. *But* we will be able to automatically generate documentation for
all flags, parameters, and command-line options, so such documentation
will always be up to date.