Add OPTIONAL_FILL_MISSING and CRITICAL_FILL_MISSING regridding flags.
Use these to replace missing values matching the _FillValue attribute
with a provided default value.

See IceModelVec::regrid() for details.

There is no way to do this from the command line, and this does not
affect user-observable behavior.

PS: This is to let hydrology models read in the sliding speed, which in
    PISM output files has missing values in ice-free areas.

PPS: In a regular run PISM should probably stop with an error message if
     a variable contains missing values. What do you think?