Fix ice discharge reporting. Port to 0.6.
We report discharge by keeping track of changes to ice
thickness (corresponds to calving off a whole column of ice) and
Href (corresponds to calving of a fraction of a column). The problem is
that changes in Href *may* also correspond to the advance of the calving
front, which is *not* associated with mass loss.

So, from now on we check if Href increased. If it did, Href's
contribution to ice discharge is zero.