Add some superficial regression tests. See #294.
We don't test correctness here; the goal is to create tests that use
enough of the PISM's Python API to be able to automatically detect
at least some changes. (I'm after good statement coverage here.)

These tests are written to be used with nose and coverage.py. See the
documenting string in nosetests.py.

I should have done this 2 (or three) years ago.