Start work coupling PISM to FEvoR.
The first step will be coupling through the spatially-variable
flow enhancement factor.

In a thermocoupled SIA setup in PISM the flow law and the enhancement
factor are used in two places:
- the computation of the SIA diffusive flux in SIAFD, and
- the computation of the volumetric strain heating in StressBalance.

The nice thing is that the IceFlowLaw class itself does not use the
enhancement factor (it *provides* it to the rest of PISM), so at this
stage we don't need to modify IceFlowLaw (or derive from it).

I created two derived classes: SIAFD_FEvoR and StressBalance_FEvoR. They
request the "enhancement_factor" field from the "Vars" dictionary and
use that instead of IceFlowLaw::enhancement_factor(). Most of the code
is copied-and-pasted from SIAFD and StressBalance, which I think is OK
at this point. New code is marked with comments.

NB! This compiles, but will *not* run: I still need to create a new PISM
component that will provide the "enhancement_factor" field. (I will
start by using a constant value so that we can compare the new setup to
the one using SIAFD and ShallowStressBalance.)