Add IceModelVec::AccessList.
This class calls begin_access() (and end_access() at the end of the
scope) for IceModelVecs that it manages. (Nothing important, but the
code is a bit cleaner.)

Note: IceModelVec::AccessList does not check the return value of
begin_access() / end_access(). This is OK: there methods should (and
will, in the future) throw an exception if a PETSc call fails.