Use the same "wide stencil" everywhere.
Added the "grid_max_stencil_width" configuration parameter (default is
2, "-calving eigen_calving" sets it to 3). IceGrid::max_stencil_width is
gone.

This commit also adds "assert(...)" macros in all the places we update
ghosts locally. This should help with making sure that all the fields
that are used have enough ghosts.

This should fix the crashing with "-calving eigen_calving" (as of the
previous commit).