Work on supporting PETSc 3.5.
All tests pass with PETSc 3.4, all but one with 3.5. This includes
Python stuff. (The bed roughness test is affected by a bug in the code
gathering IceModelVecs on processor zero; I need to fix this.)

I propose dropping the PETSc 3.3 support in the dev branch (3.4 was
released on May 22, 2013) and *requiring* PETSc 3.5 for TAO-based
inversion tools.

PISM 0.6.x should still support PETSc 3.3, 3.4, as well as 3.5 (for new
users).