if, for, while, else formatting
A conditional should always be on a line by itself, followed by {...}
UNLESS it makes the code harder to read (see SSAFD::assemble_matrix())

Same for "for" and "while".

The else keyword should be on the same line with a closing brace:
} else {...

This is a very boring change, but it should help me in the long run.

See http://google-styleguide.googlecode.com/svn/trunk/cppguide.html#Conditionals