Start work on ColumnInterpolation.
This class will contain details of the fine equally-spaced grid used for
column-wise FD computations within the ice. Currently these details are
scattered across IceGrid and iceModelVec3.

We support *two* kinds of interpolation from the coarse grid to the fine
computational grid: linear and quadratic. Only linear interpolation is
supported in the other direction, i.e. fine -> coarse.

Note that all we *really* need is the quadratic case. The linear coarse
-> fine grid interpolation was needed (a long time ago) too, because
back then the computational grid was different from the storage grid,
even if the storage grid was uniform. (Recall that we used to build
column systems for both ice and bedrock, with non-matching storage grid
spacings in ice and bedrock...) These days computational and storage
grids *match* in the uniform-storage-grid case, so no interpolation is
needed.