Add the lock() method to pism::Vars.
The idea is separate Vars initialization from its use.

When pism::Vars is created, it is empty and unlocked. During its
initialization IceModel will pass a pointer (or a reference) to
pism::Vars to sub-models which are then free to add more fields to this
dictionary, but cannot yet get pointers to fields. (This is to prevent
dependence on the initialization order of sub-models.)

Once IceModel is done allocating sub-models it locks pism::Vars, and
from this moment on one retrieve pointers to fields but cannot add new
fields or remove existing ones.