Turn pism::Vars::add() into const methods.
PISM components should be able to add new fields to this dictionary even
when pism::Vars is accessed through a const object (i.e. IceGrid). (The
grid is const because we don't want components to change grid
parameters.)

At this stage I'm not sure this is a good idea, but it helps with the
issue at hand.