Side-stepping a chicken-and-egg problem.
The problem is that sub-models may need fields provided by other
sub-models to initialize. If a sub-model A needs a field from a
sub-model B and vice versa we can't initialize both of them the same
way.

In particular, a bed deformation model may need ice thickness, so
BedDef::init() should be called after IceModel::set_vars_from_options(), but
...::set_vars_from_options() may need to override values set in
BedDef::init(), so...