Improve error propagation between C++ and Python callbacks.
A Python exception in a callback (i.e. Python code called from C++)
causes a std::exception on the C++ side, which is not caught by the C++
PISM code and returns to Python. Now we print the traceback before
handing this exception to C++ so that it is not lost once Python stops
and prints a traceback due to an uncaught exception.