Use an STL deque to manage delayed writes.
We use push_back() to add delayed writes and pop_front() when executing
them. This means that operations are performed in the order they were
requested.

Most of the time this does not matter, but a delayed write may extend an
unlimited dimension, and this has to happen *before* a new record is
written (i.e. order matters).

See also #319.