Fix chunking settings in the HDF5 I/O backend.
As of this commit "-o_format hdf5" seems to work again.

Note, though, that this code is *very* experimental and is not actively
maintained. It is kept as a starting point for somebody who might be
interested in implementing HDF5-only I/O in PISM.

As far as I know this code is used by exactly one person.

Use at your own risk and please do *not* treat this as a supported PISM
feature. For example, one should be very careful when using comparisons
of PISM's I/O performance with "-o_format hdf5" option and "-o_format
netcdf4_parallel" to reason about HDF5 and NetCDF libraries and their
performance.