Added fevor slope to the txz stress in PISMFEvoR
PISMFEvoR uses txz from a diagnostic funtion for deforming the
particles. That one didn't know of the slope, yet. Now it does.

Results still seem to be pretty much garbage, but, I think there's some truth in
these changes.

@jhkennedy In what units does fevor expect which input (esp. time), what
would be typical values for the output variables? We might be taking
waay too big timesteps right now. I'm not sure, if I got that thing
right in the previous commit.