Basic parallel lagrange tracers
Still massively todo, but runs in parallel w/o segfaulting and yields
results that are not obviously flawed when looking at them with closed
eyes. :)

Fails a bunch of the pism tests, but that's expected for now. Still
needs a lot of work.

Largely based on @jhkennedy's FEvoR code. Would have taken forever to
write without that template. Praise to him and Constantine, bugs to
me...

WARNING DON'T TRY TO USE FOR ANYTHING BUT DEBUGGING IT, YET!
This might eat the universe or have any other (unexpected) side effect.

The tracer_index variable in the netcdf is practically useless. If you
want to trace a tracer make sure, you get the correct rank and id. The
tracer_index is just a numbering in the memory that changes every time
tracers cross cpu boundaries (or make use of the periodicity of the
grid).

All testing so far was using eismint eis2 G (basically needed something
where ice flows around a bit). Don't be surprised if other things yield
complete garbage, but please tell me, if you have a simple setup that
does so.