work on I/O, kick rank from tracers
Using the process rank in the particle ID is evil
and makes processing the output annoying. We'll have to deal with
correct numbering when generating but that's less trouble than having
two ints for the id.

use get_offset to share contribution to particle numbering and get
offset that one process can start from / add to the last number in use /
...

Currently get_offset returns 0 for rank 1 -- might be smarter to have
one global offset variable that gets used in this function, but that's
not the case, yet. MIGHT CHANGE VERY SOON! DANGEROUS BEAST!