Add scripts for the "heated slab" setup.
This is the setup in which the top surface of a slab of ice sees a cold
period, followed by a warm period, and then the cold period again. After
a long enough time we expect the ice to return to the state before the
warm period.

(We need a better name for this setup.)