Re-implement lateral till water diffusion as in Bueler and Brown, 2009.
Use -hydrology null -hydrology_null_diffuse_till_water to turn it on.

I am not sure we want this, but I had to re-implement it to see if it
helps with sticky grid cells in the middle of a sliding area, for
example.

This code can (and probably should) be moved into a separate class
derived from hydrology::NullTransport.