Update run_stats in -extra_file, -save_file, -ts_file. See #330.
run_stats in -extra_file and -save_file is updated every time these
files are updated. -ts_file is tricky: we write to it once the buffer is
full, so can't easily tell if it's time to update run_stats. On the
other hand, time-series buffers are flushed every time we write to an
-extra_file, so that's when we update run_stats in -ts_file, also.