Add Jansson to PISM's build system.
PISM will look for Jansson in all standard locations. If it was not
found, we will try to download it from GitHub, build it, and install it
in PISM's build directory. (It is built as a static library, so we don't
need to install it elsewhere, neither do we need to clutter a user's
system with Jansson's header files and such.)