Remove BedrockSystem.hh and close #329.
See #40 for details. (I could implement a solver using backward-Euler
for time-stepping. This would have no time step restriction and a
maximum principle, but the code would be more complex, more
computationally expensive, and not any more accurate. Crank-Nicolson
would be more accurate, but would require a time-step restriction to
maintain the maximum principle. Note also that the current time-step
restriction is not severe enough to bother.)