Use ice density instead of mantle density in viscDisc().
See @ref BLK2006earth. In equation (43) the factor in front of the
integral has \rho_i, not \rho_r. The comment in the
code ("u(k)=rhoi*g*H0*R0*result") also refers to ice density, but the
code itself uses rho, which is mantle density.

This fixes a bug that was present in PISM since the very first SVN
revision, over 10 years ago (!).

I discovered this while working on #370. (Note that this fix does not
resolve the issue.)

Next step: use viscDisc() to create a verification test. I should have
done this years ago.