Fix an old unreported bug in the part_grid code.
In the old code (because of my mistake) results were affected by the
grid traversal order: we updated ice thickness "in place" and used
updated values at the next grid location to compute the threshold
thickness.

Now we make a copy of the ice thickness before the loop and use *it*
to compute the threshold.

I'm sorry it took me so long to find it. (It's nice to have a
verification test that can detect this kind of stuff.)