Implement mass_conservation_error_flux.
This replaces the old "nonneg_flux" and gives it a better
name (hopefully). (Without the "projection step", i.e. without the
non-negativity constraint, PISM's mass transport scheme is mass
conserving. The way we enforce non-negativity breaks mass
conservation, so the flux needed to preserve non-negativity tells us
the size of the conservation error.)