Add code needed to generate the list of supported diagnostics.
I got tired of updating the list of supported diagnostics and ensuring
that all the names in the manual are current.

This commit adds the command-line option -list_diagnostics_json, which
dumps the list of diagnostics in JSON, and the script
doc/list_diagnostics.py, which turns this JSON file into an Org mode
table (suitable for putting on GitHub).

It should now be easy to generate LaTeX tables with diagnostics,
too (even though I think this is a bad idea).

Note: when generating lists of diagnostics it is important to make
sure that *all* the sub-models are active, otherwise diagnostics they
are responsible for will be omitted from the list.