Fix a bug that resulted in running BedSmoother too often.
Recent changes to the code (the introduction of the class Geometry)
broke the code detecting if the bed elevation changed: PISM assumed
that the bed changed every time step and ran the BedSmoother every
time, slowing down the run. (Note that this did not affect modeling
results.)

Now StressBalanceInputs contains the boolean "new_bed_elevation".
SIAFD runs the BedSmoother if it is "true" and skips the BedSmoother
update if it is "false."

IceModel now has a data member "m_new_bed_elevation," set to "true"
when IceModel is created and then modified in IceModel::step().