Re-arrange mass transport code.
Now we perform a step of the mass continuity equation *without* source terms, apply
calving parameterizations, then update ocean and surface models, and then evolve ice
thickness using the source terms of the mass continuity equation. This fixes #74 and
ensures that surface and ocean models "see" the ice geometry that their output is applied
to.