Use Beckmann-Goosse everywhere instead of tracking "missing" cells.
The current version uses the Beckmann-Goosse parameterization to set "reasonable" values
at all cells containing floating ice. This way all the cells not "touched" are by the code
iterating over "boxes" are already taken care of and we don't need to keep track which
cells need this special treatment.

Also:

- instead of modifying the box mask to mark cells as "missing", remember which shelf did
  not have inputs from the previous box, telling the next iteration over boxes to use the
  Beckmann-Goosse parameterization for this shelf.

- fix spelling: "Goosse" instead of "Goose".