Fix #408
This commit changes the definition of "lakes" in the context of the PICO model.

Before this change we assumed that the biggest "ocean" (floating ice or ice-free with the
bed below sea level) area is the open ocean and all the rest are "lakes".

Now we assume that "ocean" cells that are at the edges of the computational domain belong
to the open ocean. All other ocean/floating areas (not connected to ocean points at the
edges) are "lakes."

These approaches are equivalent in Antarctic simulations, where the ice sheet is
surrounded by the ocean. The new approach works in the synthetic geometry setup referenced
in #408.