Improve the regression test
Both models (Python and PISM) have issues that need to be resolved before the PISM
implementation is ready to use.

Specifically: results are sensitive to the grid resolution. Even though we don't have an
exact solution we need to check that the code shows something resembling convergence as
the grid is refined.