Use the same ice thickness everywhere in DischargeTestingTest
The code comparing PISM's results to the value computed on the Python side samples only
one grid point: (0, 0). This works in the other test (DischargeRoutingTest) because input
fields are (intentionally) initialized so that it is at the grounding terminus.

In DischargeTestingTest, on the other hand, the point tested should be grounded with a
positive ice thickness.