Add Hydrology::flux(), remove Hydrology::velocity()
Now Hydrology has storage for the advective flux of the subglacial water (moved from
Routing).

The storage for the staggered-grid water velocity is moved back to Routing from Hydrology.

We probably want to compute and report time-averaged fluxes instead of this "snapshot"
from the last of the short time steps taken by hydrology models during a longer time step
requested by the rest of PISM.

Also: the frontal melt code needs to convert this flux from m^2/s to the corresponding
movement of the ice front (m/s).