Add a new flag: hydrology.routing.add_water_input_to_till_storage
Default: yes.

If "yes", surface water input is added to till storage which (if it overflows) then
contributes to the amount of transportable subglacial water. If "no", surface water input
directly contributes to the amount of transportable water, bypassing till storage. Basal
melt rate always contributes to the amount of water stored in the