Set WORKING_DIRECTORY for tests using nose
This is needed so that PISM can find the configuration file located in the build directory
and tests pass after "make && make test" (instead of "make install && make test").

Also: remove examples/python/bed_deformation.py from the test suite. It is not very useful
because it just tests if the code runs without checking the output.