Fix initialization of the Lingle-Clark bed deformation model
Now that the code computing the elastic response is fixed and the elastic model is enabled
by default it became clear that the elastic model shared initialization (and re-starting)
issues of the viscous model. (Related issue: #370 "Lingle-Clark bed deformation model:
stopping and re-starting affects results".)

The fix is simple: the elastic bed displacement should be treated as a part of the model
state.

Minor changes:

- Fix a regression test (viscous part of the LC bed deformation model)

  This test should run the *viscous* part of the LC model. Now that the elastic part is
  "on" by default we need to disable it in this test.

- Add elastic bed displacement to diagnostics

- Rename two regression tests

  Now all regression test of the Lingle-Clark model have "LC:" in their names. This makes
  it easier to run them using "ctest -R".