Re-write initialization of yield stress models
Yield stress models have to be able to provide basal_material_yield_stress() right after
initialization (before the first update() call) because in IceModel::step() yield stress
is used before we can compute the maximum allowed time step and so before we can call
update().

This means that "tauc" has to be a part of the model state.

Also: put MohrCoulombPointwise into a separate file.